tests(infra): remove unused infra
