--- src/tests/tty_inputs.rs
@@ -1,550 +1,0 @@
-pub const COL_141: [&str; 20] = [
-    "line1-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line2-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line3-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line4-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line5-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line6-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line7-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line8-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line9-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line10-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line11-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line12-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line13-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line14-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line15-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line16-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line17-baaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line18-baaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line19-baaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "prompt $ ",
-];
-
-pub const COL_121: [&str; 20] = [
-    "line1-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line2-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line3-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line4-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line5-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line6-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line7-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line8-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line9-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line10-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line11-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line12-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line13-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line14-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line15-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line16-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line17-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line18-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "line19-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\r\n",
-    "prompt $ ",
-];
-
-pub const COL_4: [&str; 20] = [
-    "lin1\r\n",
-    "lin2\r\n",
-    "lin3\r\n",
-    "lin4\r\n",
-    "lin5\r\n",
-    "lin6\r\n",
-    "lin7\r\n",
-    "lin8\r\n",
-    "lin9\r\n",
-    "ln10\r\n",
-    "ln11\r\n",
-    "ln12\r\n",
-    "ln13\r\n",
-    "ln14\r\n",
-    "ln15\r\n",
-    "ln16\r\n",
-    "ln17\r\n",
-    "ln18\r\n",
-    "ln19\r\n",
-    "prompt $ ",
-];
-
-pub const COL_8: [&str; 20] = [
-    "line1-bb\r\n",
-    "line2-bb\r\n",
-    "line3-bb\r\n",
-    "line4-bb\r\n",
-    "line5-bb\r\n",
-    "line6-bb\r\n",
-    "line7-bb\r\n",
-    "line8-bb\r\n",
-    "line9-bb\r\n",
-    "line10-b\r\n",
-    "line11-b\r\n",
-    "line12-b\r\n",
-    "line13-b\r\n",
-    "line14-b\r\n",
-    "line15-b\r\n",
-    "line16-b\r\n",
-    "line17-b\r\n",
-    "line18-b\r\n",
-    "line19-b\r\n",
-    "prompt $ ",
-];
-
-pub const COL_9: [&str; 20] = [
-    "line1-bbb\r\n",
-    "line2-bbb\r\n",
-    "line3-bbb\r\n",
-    "line4-bbb\r\n",
-    "line5-bbb\r\n",
-    "line6-bbb\r\n",
-    "line7-bbb\r\n",
-    "line8-bbb\r\n",
-    "line9-bbb\r\n",
-    "line10-bb\r\n",
-    "line11-bb\r\n",
-    "line12-bb\r\n",
-    "line13-bb\r\n",
-    "line14-bb\r\n",
-    "line15-bb\r\n",
-    "line16-bb\r\n",
-    "line17-bb\r\n",
-    "line18-bb\r\n",
-    "line19-bb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_10: [&str; 20] = [
-    "line1-bbbb\r\n",
-    "line2-bbbb\r\n",
-    "line3-bbbb\r\n",
-    "line4-bbbb\r\n",
-    "line5-bbbb\r\n",
-    "line6-bbbb\r\n",
-    "line7-bbbb\r\n",
-    "line8-bbbb\r\n",
-    "line9-bbbb\r\n",
-    "line10-bbb\r\n",
-    "line11-bbb\r\n",
-    "line12-bbb\r\n",
-    "line13-bbb\r\n",
-    "line14-bbb\r\n",
-    "line15-bbb\r\n",
-    "line16-bbb\r\n",
-    "line17-bbb\r\n",
-    "line18-bbb\r\n",
-    "line19-bbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_14: [&str; 20] = [
-    "line1-bbbbbbbb\r\n",
-    "line2-bbbbbbbb\r\n",
-    "line3-bbbbbbbb\r\n",
-    "line4-bbbbbbbb\r\n",
-    "line5-bbbbbbbb\r\n",
-    "line6-bbbbbbbb\r\n",
-    "line7-bbbbbbbb\r\n",
-    "line8-bbbbbbbb\r\n",
-    "line9-bbbbbbbb\r\n",
-    "line10-bbbbbbb\r\n",
-    "line11-bbbbbbb\r\n",
-    "line12-bbbbbbb\r\n",
-    "line13-bbbbbbb\r\n",
-    "line14-bbbbbbb\r\n",
-    "line15-bbbbbbb\r\n",
-    "line16-bbbbbbb\r\n",
-    "line17-bbbbbbb\r\n",
-    "line18-bbbbbbb\r\n",
-    "line19-bbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_15: [&str; 20] = [
-    "line1-bbbbbbbbb\r\n",
-    "line2-bbbbbbbbb\r\n",
-    "line3-bbbbbbbbb\r\n",
-    "line4-bbbbbbbbb\r\n",
-    "line5-bbbbbbbbb\r\n",
-    "line6-bbbbbbbbb\r\n",
-    "line7-bbbbbbbbb\r\n",
-    "line8-bbbbbbbbb\r\n",
-    "line9-bbbbbbbbb\r\n",
-    "line10-bbbbbbbb\r\n",
-    "line11-bbbbbbbb\r\n",
-    "line12-bbbbbbbb\r\n",
-    "line13-bbbbbbbb\r\n",
-    "line14-bbbbbbbb\r\n",
-    "line15-bbbbbbbb\r\n",
-    "line16-bbbbbbbb\r\n",
-    "line17-bbbbbbbb\r\n",
-    "line18-bbbbbbbb\r\n",
-    "line19-bbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_19: [&str; 20] = [
-    "line1-bbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_20: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_24: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_25: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-pub const COL_29: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_30: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_34: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_40: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_39: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_47: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_50: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_60: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_70: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_80: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_90: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
-
-pub const COL_96: [&str; 20] = [
-    "line1-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line2-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line3-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line4-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line5-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line6-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line7-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line8-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line9-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line10-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line11-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line12-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line13-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line14-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line15-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line16-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line17-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line18-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "line19-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\r\n",
-    "prompt $ ",
-];
